refactor(modals): flatten directory
