+++ packages/bezier-react/src/components/ConfirmModal/ConfirmModal.stories.tsx
