+++ packages/bezier-react/src/index.ts
@@ -29,6 +29,7 @@ export * from '~/src/components/Banner'
 export * from '~/src/components/Button'
 export * from '~/src/components/ButtonGroup'
 export * from '~/src/components/Center'
+export * from '~/src/components/ConfirmModal'
 export * from '~/src/components/Divider'
 export * from '~/src/components/Emoji'
 export * from '~/src/components/Forms'
@@ -49,7 +50,7 @@ export * from '~/src/components/Help'
 export * from '~/src/components/Icon'
 export * from '~/src/components/KeyValueListItem'
 export * from '~/src/components/ListItem'
-export * from '~/src/components/Modals'
+export * from '~/src/components/Modal'
 export * from '~/src/components/Navigator/NavGroup'
 export * from '~/src/components/Navigator/NavItem'
 export * from '~/src/components/OutlineItem'
